Fix min-height for "middle" div
According to [1], 100vh is interpreted by Chrome/Safari as "largest
possible viewport", as if the URL bar is hidden. As a result, when the
URL bar is shown, the height of the "middle" div will exceed the
viewport height, and therefore users will need to scroll down to see the
footer.

To fix the problem, change the min-height to 100%, which is defined to
be the same height of its parent. Note that according to [2],

> If the height of the containing block is not specified explicitly
> (i.e., it depends on content height), and this element is not
> absolutely positioned, the value computes to 'auto'.

Therefore, we also need to explicitly specify "height" for html and
body ("min-height" doesn't work).

An alternative solution is to use "-webkit-fill-available" and
"-moz-available" [3]. For now, the 100% solution seems sufficient.

[1] https://developer.chrome.com/blog/url-bar-resizing/
[2] https://www.w3.org/TR/CSS21/visudet.html#propdef-height
[3] https://css-tricks.com/css-fix-for-100vh-in-mobile-webkit/