Avoid specifying 98% width for footer
By default the "width" property specifies the width of the content box
of the element (excluding the padding). Since
"container_of_timer_footer" has a non-zero "padding-bottom", I guess the
"width" is set to 98% to avoid horizontal overflow. The right way to
solve the problem is to set "box-sizing" as "border-box", so that the
"width" will include the padding and border.

Remove the bottom margin of "menu" div so that the buttons won't be too
far from it when it's shown.