Allow focusing the close button for modal images
Currently the close button doesn't take focus. Make it focusable by
wrapping it with a "button" element. Although the "tabindex" attribute
could be used to control focusing order for keyboard navigation (using
TAB key), [1] suggests we write the document with the elements in a
logical sequence. Therefore, move the close button to the beginning of
the "myModal" div, so that it would become the first focusable element.
Also adjust the font size and paddings of the navigation buttons.

In addition, handle the onclick event for the modal to unset focus for
the currently active element. When the modal is open, set the visibility
of the "middle" div to hidden, so that the buttons in it won't be under
focus.

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/tabindex