Fix stacking order of the "middle" div
The z-index of the "middle" div has no effect because of its "position"
being "static" by default. Since the "background" is a positioned
element (with "fixed" position), we have to specify a positive z-index
for "middle" in order to put it on top of the background.

Therefore, change the "middle" div to "static" position.

Note that when "prefers-reduced-motion" is not set (by default), a
stacking context will be created for the "middle" div because of the
"transform" property. As a result, there will be no display issue.
However, when "prefers-reduced-motion" is enabled, animate.css will
disable the animations, and therefore the elements within "middle" might
be hidden by the background.