Fix position of floor buttons
After rotating the floor buttons ("#about", "#work"), they are no longer
positioned at the left and right borders of the window.

Fix the position by adding a translateY transform after rotation. The
translation amount is based on the newly added CSS variable, "--radius",
which is the radius of the half-circle-shaped button.